+++ partiql-cli/src/main/kotlin/org/partiql/cli/utils/Exceptions.kt
@@ -1,0 +1,11 @@
+package org.partiql.cli.utils
+
+import org.partiql.types.PartiQLValueType
+
+/** Throw an error when the there's a type mismatch when converting PartiQLValue to ExprValue */
+internal class PartiQLtoExprValueTypeMismatchException(expectedType: String, actualType: PartiQLValueType) :
+    Exception("When converting PartiQLValue to ExprValue, expected a $expectedType, but received $actualType")
+
+/** Throw an error when the there's a type mismatch when converting ExprValue to PartiQLValue */
+internal class ExprToPartiQLValueTypeMismatchException(expectedType: PartiQLValueType, actualType: PartiQLValueType) :
+    Exception("When converting ExprValue to PartiQLValue, expected a $expectedType, but received $actualType")