+++ partiql-lang/src/main/kotlin/org/partiql/lang/compiler/PartiQLCompilerDefault.kt
@@ -39,10 +39,22 @@ import org.partiql.lang.types.TypedOpParameter
 internal class PartiQLCompilerDefault(
     private val evaluatorOptions: EvaluatorOptions,
     private val customTypedOpParameters: Map<String, TypedOpParameter>,
-    private val functions: Map<String, ExprFunction>,
+    private val functions: List<ExprFunction>,
     private val procedures: Map<String, StoredProcedure>,
     private val operatorFactories: Map<RelationalOperatorFactoryKey, RelationalOperatorFactory>
 ) : PartiQLCompiler {
+
+    @Deprecated(
+        message = "functions should be a list. Use PartiQLCompilerDefault with List<ExprFunction> instead.",
+        level = DeprecationLevel.WARNING
+    )
+    constructor(
+        evaluatorOptions: EvaluatorOptions,
+        customTypedOpParameters: Map<String, TypedOpParameter>,
+        functions: Map<String, ExprFunction>,
+        procedures: Map<String, StoredProcedure>,
+        operatorFactories: Map<RelationalOperatorFactoryKey, RelationalOperatorFactory>
+    ) : this(evaluatorOptions, customTypedOpParameters, functions.values.toList(), procedures, operatorFactories)
 
     private lateinit var exprConverter: PhysicalPlanCompilerImpl
     private val bexprConverter = PhysicalBexprToThunkConverter(