+++ partiql-lang/src/main/kotlin/org/partiql/lang/eval/ExprFunction.kt
@@ -17,10 +17,8 @@ package org.partiql.lang.eval
 import org.partiql.lang.errors.ErrorCode
 import org.partiql.lang.types.FunctionSignature
 
+@Deprecated("As function overloading is now supported, this class is deprecated.", level = DeprecationLevel.ERROR)
 sealed class Arguments
-data class RequiredArgs(val required: List<ExprValue>) : Arguments()
-data class RequiredWithOptional(val required: List<ExprValue>, val opt: ExprValue) : Arguments()
-data class RequiredWithVariadic(val required: List<ExprValue>, val variadic: List<ExprValue>) : Arguments()
 
 /**
  * Represents a function that can be invoked from within an [EvaluatingCompiler]
@@ -48,39 +46,16 @@ interface ExprFunction {
      * @param required The required arguments.
      */
     fun callWithRequired(session: EvaluationSession, required: List<ExprValue>): ExprValue {
-        // Deriving ExprFunctions must implement this if they have a valid call form including only required parameters
         errNoContext("Invalid implementation for ${signature.name}#call", ErrorCode.INTERNAL_ERROR, true)
     }
 
-    /**
-     * Invokes the function with its required parameters and an optional parameter.
-     *
-     * It is assumed that the [ExprFunction]s will always be called after validating
-     * the arguments for the `arity` and the types of arguments.
-     * [EvaluatingCompiler] validates the arguments before calling [ExprFunction]s.
-     * Hence the implementations are not required to deal with it.
-     *
-     * @param session The current [EvaluationSession].
-     * @param required The required arguments.
-     * @param opt The optional arguments.
-     */
+    @Deprecated("As function overloading is now supported, please provide multiple implementations of ExprFunction and invoke callWithRequired() instead.", level = DeprecationLevel.ERROR)
     fun callWithOptional(session: EvaluationSession, required: List<ExprValue>, opt: ExprValue): ExprValue {
         // Deriving ExprFunctions must implement this if they have a valid call form including required parameters and optional
         errNoContext("Invalid implementation for ${signature.name}#call", ErrorCode.INTERNAL_ERROR, true)
     }
 
-    /**
-     * Invokes the function with its required parameters and any variadic parameters.
-     *
-     * It is assumed that the [ExprFunction]s will always be called after validating
-     * the arguments for the `arity` and the types of arguments.
-     * [EvaluatingCompiler] validates the arguments before calling [ExprFunction]s.
-     * Hence the implementations are not required to deal with it.
-     *
-     * @param session The current [EvaluationSession].
-     * @param required The required arguments.
-     * @param variadic The variadic arguments.
-     */
+    @Deprecated("As function overloading is now supported, please use the LIST ExprValue to represent variadic parameters and invoke callWithRequired() instead.", level = DeprecationLevel.ERROR)
     fun callWithVariadic(session: EvaluationSession, required: List<ExprValue>, variadic: List<ExprValue>): ExprValue {
         // Deriving ExprFunctions must implement this if they have a valid call form including required parameters and variadic
         errNoContext("Invalid implementation for ${signature.name}#call", ErrorCode.INTERNAL_ERROR, true)
@@ -98,9 +73,4 @@ interface ExprFunction {
  * @param session The current [EvaluationSession].
  * @param args The argument list.
  */
-fun ExprFunction.call(session: EvaluationSession, args: Arguments): ExprValue =
-    when (args) {
-        is RequiredArgs -> callWithRequired(session, args.required)
-        is RequiredWithOptional -> callWithOptional(session, args.required, args.opt)
-        is RequiredWithVariadic -> callWithVariadic(session, args.required, args.variadic)
-    }
+fun ExprFunction.call(session: EvaluationSession, args: List<ExprValue>): ExprValue = callWithRequired(session, args)