+++ partiql-lang/src/test/kotlin/org/partiql/lang/eval/visitors/StaticTypeInferenceVisitorTransformTest.kt
@@ -20,7 +20,6 @@ import org.partiql.lang.eval.ExprFunction
 import org.partiql.lang.eval.numberValue
 import org.partiql.lang.types.FunctionSignature
 import org.partiql.lang.types.TypedOpParameter
-import org.partiql.lang.types.VarargFormalParameter
 import org.partiql.lang.util.cartesianProduct
 import org.partiql.lang.util.compareTo
 import org.partiql.lang.util.countMatchingSubstrings
@@ -165,7 +164,7 @@ class StaticTypeInferenceVisitorTransformTest : VisitorTransformTestBase() {
     @MethodSource("parametersForAggFunctionTests")
     fun aggFunctionTests(tc: TestCase) = runTest(tc)
 
-    private fun runTest(tc: TestCase) {
+    private fun runTest(tc: TestCase) = forEachTarget {
         val globalBindings = Bindings.ofMap(tc.globals)
         val ion = IonSystemBuilder.standard().build()
         val inferencer = StaticTypeInferencer(
@@ -176,7 +175,7 @@ class StaticTypeInferenceVisitorTransformTest : VisitorTransformTestBase() {
 
         val defaultVisitorTransforms = basicVisitorTransforms()
         val staticTypeVisitorTransform = StaticTypeVisitorTransform(ion, globalBindings)
-        val originalStatement = parse(tc.originalSql).let {
+        val originalStatement = parser.parseAstStatement(tc.originalSql).let {
             // We always pass the query under test through all of the basic VisitorTransforms primarily because we need
             // FromSourceAliasVisitorTransform to execute first but also to help ensure the queries we're testing
             // make sense when they're all run.
@@ -7191,7 +7190,7 @@ class StaticTypeInferenceVisitorTransformTest : VisitorTransformTestBase() {
             ),
             TestCase(
                 "custom function",
-                "format('test %d %s', 1, 'a')",
+                "format('test %d %s', [1, 'a'])",
                 customFunctionSignatures = listOf(formatFunc.signature),
                 handler = expectQueryOutputType(StaticType.STRING)
             ),
@@ -7623,8 +7622,7 @@ class StaticTypeInferenceVisitorTransformTest : VisitorTransformTestBase() {
 
             override val signature = FunctionSignature(
                 name = "format",
-                requiredParameters = listOf(StaticType.STRING),
-                variadicParameter = VarargFormalParameter(StaticType.ANY, 0),
+                requiredParameters = listOf(StaticType.STRING, StaticType.LIST),
                 returnType = StaticType.STRING
             )
         }