+++ partiql-lang/src/test/kotlin/org/partiql/lang/syntax/PartiQLParserJoinTest.kt
@@ -7,6 +7,9 @@ import org.partiql.lang.domains.PartiqlAst
 import org.partiql.lang.domains.id
 
 class PartiQLParserJoinTest : PartiQLParserTestBase() {
+
+    override val targets: Array<ParserTarget> = arrayOf(ParserTarget.DEFAULT, ParserTarget.EXPERIMENTAL)
+
     private val projectX = PartiqlAst.build { projectList(projectExpr(id("x"))) }
 
     private fun PartiqlAst.Builder.selectWithOneJoin(