+++ partiql-lang/src/test/kotlin/org/partiql/lang/syntax/PartiQLParserMetaTests.kt
@@ -6,28 +6,30 @@ import org.partiql.lang.domains.PartiqlAst
 
 internal class PartiQLParserMetaTests : PartiQLParserTestBase() {
 
+    override val targets: Array<ParserTarget> = arrayOf(ParserTarget.DEFAULT, ParserTarget.EXPERIMENTAL)
+
     @Test
-    fun listParenthesized() {
+    fun listParenthesized(): Unit = forEachTarget {
         val query = "(0, 1, 2)"
-        val ast = parse(query) as PartiqlAst.Statement.Query
+        val ast = parser.parseAstStatement(query) as PartiqlAst.Statement.Query
         val list = ast.expr as PartiqlAst.Expr.List
 
         assert(list.metas.containsKey(IsListParenthesizedMeta.tag))
     }
 
     @Test
-    fun listParenthesizedNot() {
+    fun listParenthesizedNot(): Unit = forEachTarget {
         val query = "[0, 1, 2]"
-        val ast = parse(query) as PartiqlAst.Statement.Query
+        val ast = parser.parseAstStatement(query) as PartiqlAst.Statement.Query
         val list = ast.expr as PartiqlAst.Expr.List
 
         assert(list.metas.containsKey(IsListParenthesizedMeta.tag).not())
     }
 
     @Test
-    fun inListParenthesized() {
+    fun inListParenthesized(): Unit = forEachTarget {
         val query = "0 IN (0, 1, 2)"
-        val ast = parse(query) as PartiqlAst.Statement.Query
+        val ast = parser.parseAstStatement(query) as PartiqlAst.Statement.Query
         val inCollection = ast.expr as PartiqlAst.Expr.InCollection
         val list = inCollection.operands[1] as PartiqlAst.Expr.List
 
@@ -35,9 +37,9 @@ internal class PartiQLParserMetaTests : PartiQLParserTestBase() {
     }
 
     @Test
-    fun inListParenthesizedNot() {
+    fun inListParenthesizedNot(): Unit = forEachTarget {
         val query = "0 IN [0, 1, 2]"
-        val ast = parse(query) as PartiqlAst.Statement.Query
+        val ast = parser.parseAstStatement(query) as PartiqlAst.Statement.Query
         val inCollection = ast.expr as PartiqlAst.Expr.InCollection
         val list = inCollection.operands[1] as PartiqlAst.Expr.List
 
@@ -45,9 +47,9 @@ internal class PartiQLParserMetaTests : PartiQLParserTestBase() {
     }
 
     @Test
-    fun inListParenthesizedSingleElement() {
+    fun inListParenthesizedSingleElement(): Unit = forEachTarget {
         val query = "0 IN (0)"
-        val ast = parse(query) as PartiqlAst.Statement.Query
+        val ast = parser.parseAstStatement(query) as PartiqlAst.Statement.Query
         val inCollection = ast.expr as PartiqlAst.Expr.InCollection
         val list = inCollection.operands[1] as PartiqlAst.Expr.List